+++ libs/common/src/lib/tasks-automation/task-forms/llm-security/llm-security.component.ts
@@ -47,6 +47,7 @@ export class LLMSecurityComponent extends TaskRouteDirective {
     const parameters = {
       name: commonConfig.name,
       filter: commonConfig.filter,
+      filter_expression_json: commonConfig.filter_expression_json,
       llm: {},
       on: this.form.value.on === 'resources' ? TaskApplyTo.FULL_FIELD : TaskApplyTo.TEXT_BLOCKS,
       operations: [{ prompt_guard: { enabled: true, triggers: commonConfig.webhook && [commonConfig.webhook] } }],